segformer/requirements.txt
torch
torchvision
transformers
opencv-python
numpy
tqdm
Pillow